Formatting

+++ contracts/libraries/NFTTokenURIScaffold.sol
@@ -9,22 +9,17 @@ import {SafeMetadata} from './SafeMetadata.sol';
 import {Strings} from '@openzeppelin/contracts/utils/Strings.sol';
 import {NFTSVG} from './NFTSVG.sol';
 
-
 library NFTTokenURIScaffold {
-
     using SafeMetadata for IERC20;
     using Strings for uint256;
 
-    function tokenURI (
-        uint256 id, 
-        IPair pair, 
+    function tokenURI(
+        uint256 id,
+        IPair pair,
         IConvenience convenience,
         IPair.Due memory due,
-        uint maturity
+        uint256 maturity
     ) public view returns (string memory) {
-
-
-
         string memory uri = constructTokenSVG(
             id.toString(),
             pair.asset().safeSymbol(),
@@ -37,16 +32,18 @@ library NFTTokenURIScaffold {
             maturity
         );
 
-        string memory name = "Timeswap Collateralized Debt NFT";
-        
-        string memory description = "Timelord has blessed us with this holy NFT";
+        string memory name = 'Timeswap Collateralized Debt NFT';
+
+        string memory description = 'Timelord has blessed us with this holy NFT';
 
         return (constructTokenURI(name, description, uri));
-    
     }
 
-    function constructTokenURI (string memory name, string memory description, string memory imageSVG) internal pure returns (string memory) {
-
+    function constructTokenURI(
+        string memory name,
+        string memory description,
+        string memory imageSVG
+    ) internal pure returns (string memory) {
         return
             string(
                 abi.encodePacked(
@@ -69,21 +66,17 @@ library NFTTokenURIScaffold {
             );
     }
 
-
-
-
-    function constructTokenSVG (
+    function constructTokenSVG(
         string memory tokenId,
-        string memory assetSymbol, 
-        string memory assetAmount, 
-        string memory assetAddress, 
-        string memory collateralSymbol, 
-        string memory collateralAmount, 
+        string memory assetSymbol,
+        string memory assetAmount,
+        string memory assetAddress,
+        string memory collateralSymbol,
+        string memory collateralAmount,
         string memory collateralAddress,
         string memory maturityDate,
         uint256 maturityTimestamp
     ) internal view returns (string memory) {
-
         /// TODO - finalize SVG
         NFTSVG.SVGParams memory params = NFTSVG.SVGParams({
             tokenId: tokenId,
@@ -100,26 +93,26 @@ library NFTTokenURIScaffold {
         return NFTSVG.constructSVG(params);
     }
 
-    function weiToPrecisionString (uint256 weiAmt, uint256 decimal) public pure returns (string memory) {
+    function weiToPrecisionString(uint256 weiAmt, uint256 decimal) public pure returns (string memory) {
         if (decimal == 0) {
             return string(abi.encodePacked(weiAmt.toString(), '.00'));
         }
-        uint256 significantDigits = weiAmt/(10 ** decimal);
-        uint256 precisionDigits = weiAmt % (10 ** (decimal));
-        precisionDigits = precisionDigits/(10 ** (decimal - 2));
+        uint256 significantDigits = weiAmt / (10**decimal);
+        uint256 precisionDigits = weiAmt % (10**(decimal));
+        precisionDigits = precisionDigits / (10**(decimal - 2));
         return string(abi.encodePacked(significantDigits.toString(), '.', precisionDigits.toString()));
     }
 
-    function addressToString(address _addr) public pure returns(string memory) {
+    function addressToString(address _addr) public pure returns (string memory) {
         bytes memory data = abi.encodePacked(_addr);
-        bytes memory alphabet = "0123456789abcdef";
+        bytes memory alphabet = '0123456789abcdef';
 
         bytes memory str = new bytes(2 + data.length * 2);
-        str[0] = "0";
-        str[1] = "x";
-        for (uint i = 0; i < data.length; i++) {
-            str[2+i*2] = alphabet[uint(uint8(data[i] >> 4))];
-            str[3+i*2] = alphabet[uint(uint8(data[i] & 0x0f))];
+        str[0] = '0';
+        str[1] = 'x';
+        for (uint256 i = 0; i < data.length; i++) {
+            str[2 + i * 2] = alphabet[uint256(uint8(data[i] >> 4))];
+            str[3 + i * 2] = alphabet[uint256(uint8(data[i] & 0x0f))];
         }
         return string(str);
     }
